pyborot/requirements.txt
tensorflow==2.3.1
numpy==1.18.5
Flask==1.1.2
pandas==1.1.3
boto3==1.16.6
tensorflow_addons==0.11.2
nltk==3.5
mysql_connector_repackaged==0.3.1
scikit_learn==0.23.2
